requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-jazzmin==2.5.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
